requirements.txt
langchain
openai
streamlit
duckduckgo-search
python-dotenv
